added cache subcommands
