+++ src/commands/cache/rm.rs
@@ -1,0 +1,15 @@
+use clap::Parser;
+
+use crate::commands::Command;
+
+#[derive(Debug, Clone, Parser)]
+pub struct Args {
+    #[clap(from_global)]
+    json: bool,
+}
+
+impl Command for Args {
+    async fn runner(self) -> Result<(), anyhow::Error> {
+        todo!()
+    }
+}
